Lifecycle

+++ app/src/main/java/com/example/talaeeandroid/MainActivity.java
@@ -7,7 +7,7 @@ import android.util.Log;
 
 public class MainActivity extends AppCompatActivity {
 
-    int counter = 1;
+    short counter = 1;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
